getting there :-) Fix multi threading on HTTP RpcServer, rename rest back to http and split rest-annotations. Not fully functional yet
