+++ http/pom.xml
@@ -1,0 +1,61 @@
+<!--
+  ~ Copyright (c) Squaredesk GmbH and Oliver Dotzauer.
+  ~
+  ~ This program is distributed under the squaredesk open source license. See the LICENSE file
+  ~ distributed with this work for additional information regarding copyright ownership. You may also
+  ~ obtain a copy of the license at
+  ~
+  ~   https://squaredesk.ch/license/oss/LICENSE
+  -->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<artifactId>http</artifactId>
+
+	<parent>
+		<groupId>ch.squaredesk.nova</groupId>
+		<artifactId>parentpom</artifactId>
+		<version>4.0.0-SNAPSHOT</version>
+		<relativePath>../parentpom</relativePath>
+	</parent>
+
+	<dependencies>
+		<dependency>
+			<groupId>ch.squaredesk.nova</groupId>
+			<artifactId>spring-support</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>ch.squaredesk.nova</groupId>
+			<artifactId>comm</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+        <!--<dependency>-->
+            <!--<groupId>org.apache.httpcomponents</groupId>-->
+            <!--<artifactId>httpasyncclient</artifactId>-->
+            <!--<version>4.1.3</version>-->
+        <!--</dependency>-->
+        <dependency>
+            <groupId>com.squareup.okhttp3</groupId>
+            <artifactId>okhttp</artifactId>
+            <version>3.8.1</version>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.jersey.containers</groupId>
+            <artifactId>jersey-container-grizzly2-http</artifactId>
+            <version>2.25.1</version>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.jersey.test-framework.providers</groupId>
+            <artifactId>jersey-test-framework-provider-grizzly2</artifactId>
+            <version>2.25.1</version>
+            <scope>test</scope>
+        </dependency>
+        <!--<dependency>-->
+            <!--<groupId>org.glassfish.grizzly</groupId>-->
+            <!--<artifactId>grizresthttp-server</artifactId>-->
+            <!--<version>2.3.31</version>-->
+        <!--</dependency>-->
+
+    </dependencies>
+</project>
