+++ http/src/main/java/ch/squaredesk/nova/comm/http/HttpRequestMethod.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) Squaredesk GmbH and Oliver Dotzauer.
+ *
+ * This program is distributed under the squaredesk open source license. See the LICENSE file
+ * distributed with this work for additional information regarding copyright ownership. You may also
+ * obtain a copy of the license at
+ *
+ *   https://squaredesk.ch/license/oss/LICENSE
+ */
+
+package ch.squaredesk.nova.comm.http;
+
+public enum HttpRequestMethod {
+    GET, POST, PUT, DELETE
+}
