+++ rest-annotations/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>parentpom</artifactId>
+        <groupId>ch.squaredesk.nova</groupId>
+        <version>4.0.0-SNAPSHOT</version>
+        <relativePath>../parentpom/pom.xml</relativePath>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>rest-annotations</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>ch.squaredesk.nova</groupId>
+            <artifactId>http</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+    </dependencies>
+</project>
