+++ rest-annotations/src/main/java/ch/squaredesk/nova/comm/http/annotation/RestEnablingConfiguration.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) Squaredesk GmbH and Oliver Dotzauer.
+ *
+ * This program is distributed under the squaredesk open source license. See the LICENSE file
+ * distributed with this work for additional information regarding copyright ownership. You may also
+ * obtain a copy of the license at
+ *
+ *   https://squaredesk.ch/license/oss/LICENSE
+ */
+
+package ch.squaredesk.nova.comm.http.annotation;
+
+import ch.squaredesk.nova.Nova;
+import ch.squaredesk.nova.comm.http.HttpServerFactory;
+import ch.squaredesk.nova.comm.http.HttpServerConfiguration;
+import org.glassfish.grizzly.http.server.HttpServer;
+import org.glassfish.jersey.server.ResourceConfig;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.core.env.Environment;
+
+@Configuration
+public abstract class RestEnablingConfiguration {
+    @Autowired
+    Environment environment;
+
+    @Autowired
+    Nova nova;
+
+    @Bean
+    public RestBeanPostprocessor getBeanPostProcessor() {
+        return new RestBeanPostprocessor(resourceConfig());
+    }
+
+    @Bean
+    public ResourceConfig resourceConfig() {
+        return new ResourceConfig();
+    }
+
+    @Bean
+    public HttpServerConfiguration restServerConfiguration() {
+        int restPort = environment.getProperty("NOVA.HTTP.REST.PORT", Integer.class, 8888);
+        String interfaceName = environment.getProperty("NOVA.HTTP.INTERFACE_NAME", "");
+        if ("".equals(interfaceName)) {
+            interfaceName = "localhost";
+        }
+        return new HttpServerConfiguration(
+            interfaceName,
+            restPort
+        );
+    }
+
+    @Bean
+    @Lazy // must be created after all other beans have been created (because of annotation processing)
+    public HttpServer restHttpServer() {
+        return HttpServerFactory.serverFor(restServerConfiguration(), resourceConfig());
+    }
+}
